craw_data_airflow/requirements.txt
Twisted==21.2.0
Scrapy==2.5.0
azure-storage-blob==12.8.1
psycopg2==2.8.6